Add support for BagIt

+++ addons/weko/deposit.py
@@ -7,6 +7,7 @@ import re
 import shutil
 import tempfile
 from zipfile import ZipFile
+import bagit
 
 from framework.auth import Auth
 from framework.celery_tasks import app as celery_app
@@ -73,19 +74,37 @@ def deposit_metadata(
     self, user_id, index_id, node_id, metadata_node_id,
     schema_id, file_metadatas, project_metadatas, metadata_paths, status_path, delete_after=False,
 ):
+    def update_task_state(state=None, meta=None):
+        logger.info(f'Updating task state: {state}, {meta}')
+        self.update_state(state=state, meta=meta)
+    return _deposit_metadata(
+        user_id, index_id, node_id, metadata_node_id,
+        schema_id, file_metadatas, project_metadatas, metadata_paths, status_path, delete_after=delete_after,
+        task_request_id=self.request.id,
+        update_task_state=update_task_state,
+    )
+
+def _deposit_metadata(
+    user_id, index_id, node_id, metadata_node_id,
+    schema_id, file_metadatas, project_metadatas, metadata_paths, status_path,
+    delete_after=False, delete_temp_dir_immediately=True,
+    task_request_id=None, update_task_state=None,
+):
     user = OSFUser.load(user_id)
-    logger.info(f'Deposit: {metadata_paths}, {status_path} {self.request.id}')
+    logger.info(f'Deposit: {metadata_paths}, {status_path} {task_request_id}')
     node = AbstractNode.load(node_id)
     weko_addon = node.get_addon(SHORT_NAME)
-    weko_addon.set_publish_task_id(status_path, self.request.id)
+    weko_addon.set_publish_task_id(status_path, task_request_id)
     wb = WaterButlerClient(user).get_client_for_node(node)
     tmp_dir = None
+    bagit_dir = None
     try:
         tmp_dir = tempfile.mkdtemp()
-        self.update_state(state='downloading', meta={
-            'progress': 10,
-            'paths': metadata_paths,
-        })
+        if update_task_state:
+            update_task_state(state='downloading', meta={
+                'progress': 10,
+                'paths': metadata_paths,
+            })
         download_file_names = []
         download_files = []
         total_size = 0
@@ -93,10 +112,11 @@ def deposit_metadata(
             path = metadata_path
             if '/' not in path:
                 raise ValueError(f'Malformed path: {path}')
-            self.update_state(state='initializing', meta={
-                'progress': 0,
-                'path': metadata_path,
-            })
+            if update_task_state:
+                update_task_state(state='initializing', meta={
+                    'progress': 0,
+                    'path': metadata_path,
+                })
             materialized_path = path[path.index('/'):]
             file = wb.get_file_by_materialized_path(path)
             logger.debug(f'File: {file}, size={file.size}')
@@ -109,32 +129,60 @@ def deposit_metadata(
             download_file_names.append((download_file_name, download_file_type))
             download_files.append(file)
             logger.info(f'Downloaded: {download_file_path} {filesize}')
-        self.update_state(state='packaging', meta={
-            'progress': 50,
-            'paths': metadata_paths,
-        })
+        if update_task_state:
+            update_task_state(state='packaging', meta={
+                'progress': 50,
+                'paths': metadata_paths,
+            })
 
         c = weko_addon.create_client()
         target_index = c.get_index_by_id(index_id)
 
-        # TODO SimpleZip -> SWORD BagIt
+        # Packaging the files as BagIt
+        bagit_dir = tempfile.mkdtemp()
+        bagit_metadata = {
+            'Contact-Name': user.fullname,
+            'Contact-Email': user.username,
+        }
+        if user.affiliated_institutions and user.affiliated_institutions.first():
+            bagit_metadata['Source-Organization'] = user.affiliated_institutions.first().name
+        bag = bagit.make_bag(bagit_dir, bagit_metadata)
+
+        for download_file_name, _ in download_file_names:
+            file_in_bagit_path = os.path.join(bagit_dir, 'data', 'files', download_file_name)
+            os.makedirs(os.path.dirname(file_in_bagit_path), exist_ok=True)
+            shutil.copyfile(os.path.join(tmp_dir, download_file_name), file_in_bagit_path)
+        # Metadata as CSV
+        with open(os.path.join(bagit_dir, 'data', 'index.csv'), 'w', encoding='utf8') as f:
+            schema.write_csv(
+                user,
+                f,
+                target_index,
+                download_file_names,
+                schema_id,
+                file_metadatas,
+                project_metadatas,
+            )
+        # Metadata as RO-Crate
+        with open(os.path.join(bagit_dir, 'data', 'ro-crate-metadata.json'), 'w', encoding='utf8') as f:
+            schema.write_ro_crate_json(
+                user,
+                f,
+                target_index,
+                download_file_names,
+                schema_id,
+                file_metadatas,
+                project_metadatas,
+            )
+        bag.save(manifests=True)
+
         zip_path = os.path.join(tmp_dir, 'payload.zip')
-        with ZipFile(zip_path, 'w') as zf:
-            for download_file_name, _ in download_file_names:
-                with zf.open(f'data/files/{download_file_name}', 'w') as df:
-                    with open(download_file_path, 'rb') as sf:
-                        shutil.copyfileobj(sf, df)
-            with zf.open('data/index.csv', 'w') as f:
-                with io.TextIOWrapper(f, encoding='utf8') as tf:
-                    schema.write_csv(
-                        user,
-                        tf,
-                        target_index,
-                        download_file_names,
-                        schema_id,
-                        file_metadatas,
-                        project_metadatas,
-                    )
+        with ZipFile(zip_path, 'w') as zipf:
+            for root, dirs, files in os.walk(bagit_dir):
+                for file in files:
+                    file_path = os.path.join(root, file)
+                    zipf.write(file_path, os.path.relpath(file_path, bagit_dir))
+
         headers = {
             'Packaging': 'http://purl.org/net/sword/3.0/package/SimpleZip',
             'Content-Disposition': 'attachment; filename=payload.zip',
@@ -142,17 +190,19 @@ def deposit_metadata(
         files = {
             'file': ('payload.zip', open(zip_path, 'rb'), 'application/zip'),
         }
-        self.update_state(state='uploading', meta={
-            'progress': 60,
-            'paths': metadata_paths,
-        })
+        if update_task_state:
+            update_task_state(state='uploading', meta={
+                'progress': 60,
+                'paths': metadata_paths,
+            })
         logger.info(f'Uploading... {file_metadatas}')
         respbody = c.deposit(files, headers=headers)
         logger.info(f'Uploaded: {respbody}')
-        self.update_state(state='uploaded', meta={
-            'progress': 100,
-            'paths': metadata_paths,
-        })
+        if update_task_state:
+            update_task_state(state='uploaded', meta={
+                'progress': 100,
+                'paths': metadata_paths,
+            })
         links = [l for l in respbody['links'] if 'contentType' in l and '@id' in l and l['contentType'] == 'text/html']
         for file in download_files:
             if delete_after:
@@ -163,7 +213,7 @@ def deposit_metadata(
                 {
                     'materialized': file.materialized,
                     'path': file.path,
-                    'item_html_url': links[0]['@id'],
+                    'item_html_url': links[0]['@id'] if len(links) > 0 else None,
                 },
             )
         return {
@@ -171,5 +221,7 @@ def deposit_metadata(
             'paths': metadata_paths,
         }
     finally:
-        if tmp_dir and os.path.exists(tmp_dir):
+        if delete_temp_dir_immediately and tmp_dir and os.path.exists(tmp_dir):
             shutil.rmtree(tmp_dir)
+        if delete_temp_dir_immediately and bagit_dir and os.path.exists(bagit_dir):
+            shutil.rmtree(bagit_dir)
